+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Update: Improved MenuBuilder; Completed assertions; Added role handling
 
 - Added selectable options to the enums static methods trait
 - Completed route handling in the SelectOptions trait
+- Defer to the Roles enum defined in auth.providers.roles.enum config
 
 ---